remove materialize collapsible
