+++ src/sass/_notifications.scss
@@ -6,6 +6,10 @@
   top: 125px;
   margin: 0 0 0 1px;
 
+  ul {
+    margin: 0;
+  }
+
   & > div {
     width: 100%;
     height: 100%;
@@ -18,15 +22,18 @@
     border: rgba(196, 196, 196, 0.3) solid 1px;
   }
 
-  .collapsible-header, .collapsible-body {
-    padding: 0;
+  .dojot-collapsible-header {
+    cursor: pointer;
   }
 
-  .collapsible {
-    margin: 0;
-    box-shadow: none;
+  .dojot-collapsible-body {
+    display: none;
   }
-  
+
+  .dojot-collapsible-body-active {
+    display: block;
+  }
+
   .card-notification {
     display: flex;
     min-height: 70px;
@@ -58,7 +65,7 @@
       }
     }
 
-    .first-col {
+    .first-col-noti {
       background: rgba(196, 196, 196, 0.2);
       width: 150px;
       text-align: right;
@@ -93,17 +100,17 @@
       }
     }
 
-    .second-col {
+    .second-col-noti {
       font-style: normal;
       font-weight: 500;
       justify-content: space-between;
       align-items: center;
       display: flex;
       width: 100%;
-      padding: 4px 0;
+      padding: 10px 0;
 
       > div {
-        padding: 4px 6px;
+        padding: 6px 15px;
       }
